desktop-ui/pyproject.toml
[project]
name = "desktop-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "piexif>=1.1.3",
    "pillow>=11.2.1",
    "wxpython>=4.2.3",
]
